Move util out of enterprise, apply on user profile
