+++ client/web/src/enterprise/rbac/components/RoleNode.tsx
@@ -25,8 +25,8 @@ import {
 
 import { LoaderButton } from '../../../components/LoaderButton'
 import { RoleFields } from '../../../graphql-operations'
+import { prettifySystemRole } from '../../../util/settings'
 import { PermissionsMap, useDeleteRole, useSetPermissions } from '../backend'
-import { prettifySystemRole } from '../util'
 
 import { ConfirmDeleteRoleModal } from './ConfirmDeleteRoleModal'
 import { PermissionsList } from './Permissions'
