+++ client/web/src/user/area/UserArea.tsx
@@ -50,6 +50,7 @@ export const UserAreaGQLFragment = gql`
         roles @skip(if: $isSourcegraphDotCom) {
             nodes {
                 name
+                system
             }
         }
     }
